==> Final Files To Submit/R Scripts/4 Wilcoxon_Results.Rmd <==
---
title: "Wilcoxon Results Report"
author: "Joseph Oliveira"
date: "7/5/2020"
output: 
  pdf_document:
    toc: TRUE
---


```{r include=FALSE}
library(tidyverse)
library(purrr)
source('FACES_fns.R')
```

# Comparisons

Each row represents a comparison. 

```{r include=FALSE}
distinct_groupings <- read_csv('../Data/cleaned_FACES_data.csv', col_types = c('ffffdf')) %>%
  distinct(Group, Time, Survey)

dist_grp_exp_pre <- distinct_groupings %>%
  filter(Group == 'Experimental', Time == 'Pre')

dist_grp_ctrl_pre <- distinct_groupings %>%
  filter(Group == 'Control', Time == 'Pre')

dist_grp_exp_post <- distinct_groupings %>%
  filter(Group == 'Experimental', Time == 'Post')

dist_grp_ctrl_post <- distinct_groupings %>%
  filter(Group == 'Control', Time == 'Post')

first_comp <- inner_join(dist_grp_exp_pre, dist_grp_ctrl_pre, by = c("Survey", "Time"))
secnd_comp <- inner_join(dist_grp_exp_post, dist_grp_ctrl_post, by = c("Survey", "Time"))
third_comp <- inner_join(dist_grp_exp_pre, dist_grp_exp_post, by = c("Survey", "Group"))
forth_comp <- inner_join(dist_grp_ctrl_pre, dist_grp_ctrl_post, by = c("Survey", "Group"))

group_comparison <- bind_rows(first_comp, secnd_comp) %>% mutate_all(as.character) %>%
  select(Group.x, Group.y, Time, Survey)

time_comparison <- bind_rows(third_comp, forth_comp) %>% mutate_all(as.character) %>%
  select(Time.x, Time.y, Group, Survey)
```

The total of Pre and the total of Post intervention responses for the Control group is compared to the total of Pre and the total of Post intervention responses for the Experimental group. 

```{r echo=FALSE}
knitr::kable(group_comparison,
             caption = 'Experimental vs. Control Groups ')
```

For each the Experimental and the Control group, the total of Pre intervention responses were compared to the total of Post intervention responses. 

```{r echo=FALSE}
knitr::kable(time_comparison,
             caption = 'Pre vs Post within Experimental and Control Groups')
```

\pagebreak 

# Results

```{r include=FALSE}
group_wilcoxon <- group_comparison %>%
  # defined the test result
  mutate(test = pmap(list(time1 = Time, group1 = Group.x, 
                          time2 = Time, group2 = Group.y, 
                          survey = Survey), 
                     faces_wilcox, paired = F, alternative = "two.sided"),
         test_statistic = unlist(map(test, function(x) `$`(`[`(`[`(x)), 'statistic'))),
         p.value = unlist(map(test, function(x) `$`(`[`(`[`(x)), 'p.value'))),
         effect_size =  test_statistic / unlist(map(test, function(x) `$`(`[`(`[`(x)), 'obs')))**0.5)

time_wilcoxon  <- time_comparison %>%
  mutate(test = pmap(list(time1 = Time.x, group1 = Group, 
                          time2 = Time.y, group2 = Group, 
                          survey = Survey),
                     faces_wilcox, paired = T, alternative = "less"),
         test_statistic = unlist(map(test, function(x) `$`(`[`(`[`(x)), 'statistic'))),
         p.value = unlist(map(test, function(x) `$`(`[`(`[`(x)), 'p.value'))),
         effect_size = test_statistic / unlist(map(test, function(x) `$`(`[`(`[`(x)), 'obs')))**0.5)
```

## Experimental vs. Control Groups Results

```{r echo=FALSE}
knitr::kable(group_wilcoxon %>%
               filter(Time == 'Pre') %>%
               select(Time, Survey, 
                    `Group A` = Group.x, 
                    `Group B` = Group.y, 
                    `W` = test_statistic, 
                    `p` = p.value#, 
                    #`ES` = effect_size
                    ) %>%
               mutate(p = round(p, 3)#,
                      #ES = round(ES, 3)
                      ) %>%
               group_by(Survey) %>%
               arrange(Survey),
             caption = 'Pre Group Results')
```

  - $H_0$: There is no difference between the populations of *Pre Experimental* group and the *Pre Control* group.
  - $H_A$: There is a difference between the populations of *Pre Experimental* group and the *Pre Control* group.

```{r  echo=FALSE}
knitr::kable(group_wilcoxon %>%
               filter(Time == 'Post') %>%
               select(Time, Survey, 
                    `Group A` = Group.x, 
                    `Group B` = Group.y, 
                    `W` = test_statistic , 
                    `p` = p.value#, 
                    #`ES` = effect_size
                    ) %>%
               mutate(p = round(p, 3)#,
                      #ES = round(ES, 3)
                      ) %>%
               group_by(Survey) %>%
               arrange(Survey),
             caption = 'Post Group Results')
```

  - $H_0$: There is no difference between the populations of *Post Experimental* group and the *Post Control* group.
  - $H_A$: There is a difference between the populations of *Post Experimental* group and the *Post Control* group.

W  = Test Statistic from Wilcoxon Test  
p  = P-Value of W  
ES = Effect Size W

\pagebreak

## Pre vs. Post Results

```{r echo=FALSE}
knitr::kable(time_wilcoxon %>%
               filter(Group == "Experimental") %>%
               select(Group, Survey, 
                    `Pre` = Time.x, 
                    `Post` = Time.y, 
                    `W` = test_statistic, `p` = p.value
                    #, `ES` = effect_size
                    ) %>%
               mutate(p = round(p, 3)#,
                      #ES = round(ES, 3)
                      ) %>%
               group_by(Survey) %>%
               arrange(Survey),
             caption = 'Experimental Group Results')
```

  - $H_0$: There is no difference between the populations of *Pre Experimental* group and the *Post Experimental* group.
  - $H_A$: There is a difference between the populations of *Pre Experimental* group and the *Post Experimental* group.

```{r echo=FALSE}
knitr::kable(time_wilcoxon %>%
               filter(Group == "Control") %>%
               select(Group, Survey, 
                    `Pre` = Time.x, 
                    `Post` = Time.y, 
                    `W` = test_statistic, `p` = p.value
                    #, `ES` = effect_size
                    ) %>%
               mutate(p = round(p, 3),
                     # ES = round(ES, 3)
                     ) %>%
               group_by(Survey) %>%
               arrange(Survey),
             caption = 'Control Group Results')
```


  - $H_0$: There is no difference between the populations of *Pre Control* group and the *Post Control* group.
  - $H_A$: There is a difference between the populations of *Pre Control* group and the *Post Control* group.

W  = Test Statistic from Wilcoxon Test  
p  = P-Value of W   
~~ES = Effect Size W~~
