==> Final Files To Submit/R Scripts/5 CreateParticipantLevelBarCharts.RMD <==
---
title: "Data visualizations"
author: "Shantel Ward"
date: "7/8/2020"
output: pdf_document
---
  

```{r message=FALSE, warning=FALSE}
library(tidyverse)
theme_set(theme_bw())
faces <- read_csv("C:/Users/17739/Desktop/ST 542 Statistical Practice/FinalProject/FACES/Data/cleaned_FACES_data.csv")
#faces <- read_csv("../../../../Data/cleaned_FACES_data.csv")
faces_tbl <- faces %>% 
  mutate(Survey = factor(Survey)) %>%
  group_by(`Participant #`, Time, Group, Survey) %>%
  summarise(Total_Score = sum(Response),
            Total_Questions = length(unique(Question)))

faces_tbl$`Participant #`<- as.factor(faces_tbl$`Participant #`) #used factor instead of character to preserve order 


setwd("C:/Users/17739/Desktop/ST 542 Statistical Practice/FinalProject/FACES/Documents/Script_Documents/Visuals/SNW")

createBarPlot <- function(survey) {
  
  category_Exp<-faces_tbl%>%filter(Survey==survey, Group=="Experimental")

  #meanCalc<-category_Exp%>%group_by(Time)%>%summarise(Mean=mean(Total_Score))
  #extractmeanPre<-meanCalc%>%filter(Time=="Pre")%>%select(Mean)
  #meanPre<-extractmeanPre$Mean[1]
  #extractmeanPost<-meanCalc%>%filter(Time=="Post")%>%select(Mean)
  #meanPost<-extractmeanPost$Mean[1]

  medianCalc<-category_Exp%>%group_by(Time)%>%summarise(Median=median(Total_Score))
  extractmedianPre<-medianCalc%>%filter(Time=="Pre")%>%select(Median)
  medianPre<-extractmedianPre$Median[1]
  extractmedianPost<-medianCalc%>%filter(Time=="Post")%>%select(Median)
  medianPost<-extractmedianPost$Median[1]

  ggplot(category_Exp, aes(x = `Participant #`,y = Total_Score, group = Time)) + 
    geom_col(position = "dodge",aes(fill=Time), color = 'black') +
    coord_flip() +
    geom_hline(aes(yintercept = medianPre, linetype = "Median of Pre"), size=1, color = "green") +
    geom_hline(aes(yintercept = medianPost, linetype = "Median of Post"), size=1, color = "orange") +
    geom_text(aes(label =""), position = position_dodge(width = .75)) +
    scale_linetype_manual(name = "Median Lines", values = c(2, 2), 
                      guide = guide_legend(override.aes = list(color = c("orange", "green")))) +
    ggtitle(paste(survey," Scores - Experimental Group (Pre vs Post)")) +
    scale_y_continuous(breaks = scales::pretty_breaks(10)) + 
    theme(plot.title = element_text(hjust = 0.5)) #to center title 

  ggsave(plot = last_plot(), filename = paste0(survey,"_Experimental_post_vs_pre.png"), width = 6, height = 4, units = 'in')
  
}

createBarPlot("FES")
createBarPlot("AKS")
createBarPlot("FACES")
createBarPlot("FPPS")
createBarPlot("SCS")
createBarPlot("SEAS")

```
